\name{ad-class}
\Rdversion{1.1}
\docType{class}
\alias{ad}
\alias{ad-class}


\title{Class \code{"ad"}}
\description{
An S4 class for the Anderson-Darling test for multivariate normality.
}

\section{Slots}{
  \describe{
    \item{\code{AD}:}{stores the values of the test statistics}
    \item{\code{p.value}:}{stores the p-value of the test}
    \item{\code{data.name}:}{stores the name of the data}
  }
}
\author{
Natalya Pya, Vassilly Voinov, Rashid Makarov
}
