\name{cm-class}
\Rdversion{1.1}
\docType{class}
\alias{cm}
\alias{cm-class}


\title{Class \code{"cm"}}
\description{
An S4 class for the Cramer-von Mises test for multivariate normality.
}

\section{Slots}{
  \describe{
    \item{\code{CM}:}{stores the values of the test statistics}
    \item{\code{p.value}:}{stores the p-value of the test}
    \item{\code{data.name}:}{stores the name of the data}
  }
}
\author{
   Natalya Pya, Vassilly Voinov, Rashid Makarov
}
