\name{dh-class}
\Rdversion{1.1}
\docType{class}
\alias{dh}
\alias{dh-class}


\title{Class \code{"dh"}}
\description{
An S4 class for the Doornik-Hansen test for multivariate normality.
}

\section{Slots}{
  \describe{
    \item{\code{DH}:}{stores the values of the test statistics}
    \item{\code{p.value}:}{stores the p-value of the test}
    \item{\code{data.name}:}{stores the name of the data}
  }
}
\author{
   Natalya Pya, Vassilly Voinov, Rashid Makarov
}
