\name{hz-class}
\Rdversion{1.1}
\docType{class}
\alias{hz}
\alias{hz-class}


\title{Class \code{"hz"}}
\description{
An S4 class for the Henze-Zirkler test for multivariate normality.
}

\section{Slots}{
  \describe{
    \item{\code{HZ}:}{stores the values of the test statistics}
    \item{\code{p.value}:}{stores the p-value of the test}
    \item{\code{data.name}:}{stores the name of the data}
  }
}
\author{
   Natalya Pya, Vassilly Voinov, Rashid Makarov
}
