\name{mvnTest-package}
%\Rdversion{1.1-0}
\alias{mvnTest-package}
\docType{package}
\title{Goodness-of-fit tests for multivariate normality
}
\description{
\code{mvnTest} provides functions for assessing multivariate normality. It includes eight test statistics such as three Wald's type chi-squared tests (McCulloch, Nikulin-Rao-Robson and Dzhaparidze-Nikulin tests), non-parametric Anderson-Darling and Cramer-von Mises tests; Doornik-Hansen test,  Royston test and Henze-Zirkler test.
}
\details{
\tabular{ll}{
Package: \tab mvnTest\cr
Type: \tab Package\cr
License: \tab GPL (>= 2)\cr
}}
\author{
Vassilly Voinov, Natalya Pya, Rashid Makarov, Yevgeniy Voinov

Maintainer: Natalya Pya <nat.pya@gmail.com>

}
\note{
   The printing method and plotting are in part adapted from R package \code{MVN} 
(version 4.0, Korkmaz, S. et al., 2015)
} 
