\name{r-class}
\Rdversion{1.1}
\docType{class}
\alias{r}
\alias{r-class}


\title{Class \code{"r"}}
\description{
An S4 class for the Royston test for multivariate normality.
}

\section{Slots}{
  \describe{
    \item{\code{R}:}{stores the values of the test statistics}
    \item{\code{p.value}:}{stores the p-value of the test}
    \item{\code{data.name}:}{stores the name of the data}
  }
}
\author{
   Natalya Pya, Vassilly Voinov, Rashid Makarov
}
